Interleaves event handler streams to solve no delay starvation.
