+++ crate/event_loop/Cargo.toml
@@ -10,8 +10,13 @@ test = false
 
 [dependencies]
 async-std = { version = "1.6.0-beta.1", optional = true }
+futures-test = { version = "0.3.5", optional = true }
 futures = { version =  "0.3.5", default-features = false }
 
 [features]
 default = []
-rate_limit = ["async-std/unstable"]
+rate_limit = [
+    "async-std",
+    "async-std/unstable",
+    "futures-test",
+]
